requirements.txt
altgraph==0.17.2
certifi==2024.8.30
charset-normalizer==3.4.0
clipboard==0.0.4
filelock==3.16.1
fsspec==2024.10.0
future==0.18.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
macholib==1.15.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.3.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
pyvenv==0.2.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==58.0.4
six==1.15.0
sympy==1.13.1
termcolor==1.1.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.37.0
XlsxWriter==3.2.0